refactor(layout): modularize route group layouts and remove redundant root layout
